backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-decouple>=3.8
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.26
pytz
gunicorn>=21.0
whitenoise>=6.0
django-storages>=1.14
boto3>=1.28
Pillow